[fix] env_lst add key
